Add files via upload
This is really when things started to kick off for my programming skills and most of my poorly written code started to get much better, although at this time I couldn't fix many issues as it would have extended the release date for the game and I had to take a few liberties, I think I also did this in version 3. But this is version 4, is it good? maybe not, but is it better than when I started, YES